Updated to have no python_version requirement and ignore lock
From the docs, it said we can remove it and then it targets the main version installed on the system. The lock then is kept on the local machine. This allows multiple dependency packages for multiple versions of python. When we are all under the same, we can re-commit this.